Huge server refactor
