+++ server/templates/checks.html
@@ -18,17 +18,17 @@
   </tr>
 </thead>
 <tbody>
-    {% for name in names %}
+    {% for cname in clean_names %}
     <tr>
         <td>{{ loop.index }}</td>
         <td>
-            <button type="button" class="btn btn-xs- btn-default" onclick="location.href='/inspect/{{ instance }}/{{ database }}/{{ table }}/{{ name }}';">
-                {{ name }}
+            <button type="button" class="btn btn-xs- btn-default" onclick="location.href='/inspect/{{ instance }}/{{ database }}/{{ table }}/{{ names[loop.index0] }}';">
+                {{ names[loop.index0] }}
             </button>
         </td>
         <td>
             <h4>
-            {% if metadata[name]['passing'] == 0 %}
+            {% if metadata[cname]['passing'] == 0 %}
                 <span class="label label-success">
                     Passing
                 </span>
@@ -39,7 +39,7 @@
             {% endif %}
             </h4>
         </td>
-        <td>{{ metadata[name]['passing'] }}</td>
+        <td>{{ metadata[cname]['passing'] }}</td>
     </tr>
     {% endfor %}
 </tbody>
@@ -48,7 +48,7 @@
 {% block js %}
 <script>
 function convertDataWeek() {
-{% for name in names %}
+{% for name in clean_names %}
     var {{ name }} = [
     {% for row in history[name]['week'] %}
         {
@@ -60,17 +60,17 @@ function convertDataWeek() {
 {% endfor %}
 
     return [
-        {% for name in names %}
+        {% for name in clean_names %}
         {
             values: {{ name }},
             key: '{{ name }}',
             color: '{{ colors[loop.index % 9] }}'
-        } {% if loop.index != names|length %},{% endif %}
+        } {% if loop.index != clean_names|length %},{% endif %}
         {% endfor %}
     ];
 };
 function convertDataMonth() {
-{% for name in names %}
+{% for name in clean_names %}
     var {{ name }} = [
     {% for row in history[name]['month'] %}
         {
@@ -82,17 +82,17 @@ function convertDataMonth() {
 {% endfor %}
 
     return [
-        {% for name in names %}
+        {% for name in clean_names %}
         {
             values: {{ name }},
             key: '{{ name }}',
             color: '{{ colors[loop.index % 9] }}'
-        } {% if loop.index != names|length %},{% endif %}
+        } {% if loop.index != clean_names|length %},{% endif %}
         {% endfor %}
     ];
 };
 function convertDataYear() {
-{% for name in names %}
+{% for name in clean_names %}
     var {{ name }} = [
     {% for row in history[name]['year'] %}
         {
@@ -104,12 +104,12 @@ function convertDataYear() {
 {% endfor %}
 
     return [
-        {% for name in names %}
+        {% for name in clean_names %}
         {
             values: {{ name }},
             key: '{{ name }}',
             color: '{{ colors[loop.index % 9] }}'
-        } {% if loop.index != names|length %},{% endif %}
+        } {% if loop.index != clean_names|length %},{% endif %}
         {% endfor %}
     ];
 };
